# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Booking API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/bookings:
    get:
      operationId: getMyBookings
      tags: [users]
      summary: Get booking history for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of user bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingWithDetails"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Admin: Users
  /admin/users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Hotels
...
  /hotels/{hotelId}/rooms:
    get:
      operationId: getRoomsByHotel
      tags: [rooms]
      summary: Get all rooms for a hotel
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
        - name: checkIn
          in: query
          schema:
            type: string
        - name: checkOut
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of rooms with availability
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomWithAvailability"

    post:
      operationId: createRoom
      tags: [rooms]
      summary: Add a room to a hotel (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomBody"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}:
    patch:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoomBody"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Room deleted

  # Bookings
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    Hotel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        rating:
          type: number
        stars:
          type: integer
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        reviewCount:
          type: integer
        minPrice:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - city
        - address
        - rating
        - stars
        - amenities
        - images
        - reviewCount
        - createdAt
...
    Room:
      type: object
      properties:
        id:
          type: integer
        hotelId:
          type: integer
        type:
          type: string
          enum: [single, double, deluxe, suite]
        price:
          type: number
        guests:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - hotelId
        - type
        - price
        - guests
        - description
        - images
        - createdAt
    RoomWithAvailability:
      type: object
      properties:
        id:
          type: integer
        hotelId:
          type: integer
        type:
          type: string
          enum: [single, double, deluxe, suite]
        price:
          type: number
        guests:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        isAvailable:
          type: boolean
      required:
        - id
        - hotelId
        - type
        - price
        - guests
        - description
        - images
        - createdAt
        - isAvailable
    CreateRoomBody:
      type: object
      properties:
        type:
          type: string
          enum: [single, double, deluxe, suite]
        price:
          type: number
        guests:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
      required:
        - type
        - price
        - guests
        - description
    UpdateRoomBody:
      type: object
      properties:
        type:
          type: string
          enum: [single, double, deluxe, suite]
        price:
          type: number
        guests:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
...
    BookingWithDetails:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        roomId:
          type: integer
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        createdAt:
          type: string
          format: date-time
        room:
          $ref: "#/components/schemas/Room"
        hotel:
          $ref: "#/components/schemas/Hotel"
      required:
        - id
        - userId
        - roomId
        - checkIn
        - checkOut
        - totalPrice
        - status
        - createdAt
        - room
        - hotel
...